requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
jinja2==3.1.5
python-multipart==0.0.20
numpy==1.26.4
faiss-cpu==1.10.0
sentence-transformers==3.0.1
